hwtest-nats/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "hwtest-nats"
version = "0.1.0"
description = "NATS/JetStream integration for hwtest telemetry streaming"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "hwtest contributors"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "hwtest-core",
    "nats-py>=2.6",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "black>=23.0",
    "mypy>=1.0",
    "pylint>=3.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
markers = [
    "integration: marks tests requiring NATS server (deselect with '-m \"not integration\"')",
]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.mypy]
strict = true
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true

[tool.pylint.main]
max-line-length = 100
disable = [
    "too-few-public-methods",
]
